transdata_op.cpp&op_executor.cpp:misra fix

+++ single_op/op_executor.cpp
@@ -186,7 +186,7 @@ aclError OpExecutor::DoExecuteAsync(ge::DynamicSingleOp *singleOp,
 
     if (aclOp.exeucteType == ACL_OP_EXECUTE_V2) {
         ACL_LOG_INFO("Begin to update outputDesc");
-        for (size_t i = 0; i < outputDesc.size(); ++i) {
+        for (size_t i = 0U; i < outputDesc.size(); ++i) {
             ge::GeShape outputShape = outputDesc[i].GetShape();
             std::vector<int64_t> outputDims = outputShape.GetDims();
             ACL_LOG_INFO("update outputDesc[%zu] dims is [%s]", i, string_utils::VectorToString(outputDims).c_str());
@@ -315,7 +315,7 @@ aclError OpExecutor::ExecuteAsync(OpHandle &opHandle,
                 }
 
                 // Just for protection. Preventing from cache grows to large.
-                if (cachedExecutors.size() > MAX_CACHED_NUM) {
+                if (cachedExecutors.size() > static_cast<size_t>(MAX_CACHED_NUM)) {
                     auto toErase = cachedExecutors.begin();
                     ACL_LOG_WARN("cache[%zu] reaches max size[%zu], evict one object. stream = %p",
                         cachedExecutors.size(), MAX_CACHED_NUM, toErase->first);
@@ -344,7 +344,7 @@ aclError OpExecutor::ExecuteAsync(OpHandle &opHandle,
                 }
 
                 // Just for protection. Preventing from cache grows to large.
-                if (cachedExecutors.size() > MAX_CACHED_NUM) {
+                if (cachedExecutors.size() > static_cast<size_t>(MAX_CACHED_NUM)) {
                     auto toErase = cachedExecutors.begin();
                     ACL_LOG_INFO("cache[%zu] reaches max size[%zu], evict one object. stream = %p",
                         cachedExecutors.size(), MAX_CACHED_NUM, toErase->first);
